pyproject.toml
[project]
name = "daida-ai"
version = "0.1.0"
description = "LT presentation auto-generation Agent Skill"
requires-python = ">=3.11"
dependencies = [
    "python-pptx>=1.0.0",
    "edge-tts>=6.1.0",
    "pydub>=0.25.1",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
voicevox = ["httpx>=0.27.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
